+++ src/client/scilexer/PositionCache.cpp
@@ -606,11 +606,11 @@ void PositionCache::MeasureWidths(Surface *surface, ViewStyle &vstyle, unsigned 
 
 		// Two way associative: try two probe positions.
 		int hashValue = PositionCacheEntry::Hash(styleNumber, s, len);
-		probe = hashValue % size;
+		probe = hashValue % (int)size;
 		if (pces[probe].Retrieve(styleNumber, s, len, positions)) {
 			return;
 		}
-		int probe2 = (hashValue * 37) % size;
+		int probe2 = (hashValue * 37) % (int)size;
 		if (pces[probe2].Retrieve(styleNumber, s, len, positions)) {
 			return;
 		}
