+++ src/client/windows/nxcon/UserSelectDlg.cpp
@@ -124,7 +124,7 @@ void CUserSelectDlg::OnOK()
       iItem = m_wndListCtrl.GetNextItem(-1, LVIS_SELECTED);
       for(i = 0; (iItem != -1) && (i < m_dwNumUsers); i++)
       {
-         m_pdwUserList[i] = m_wndListCtrl.GetItemData(iItem);
+         m_pdwUserList[i] = (DWORD)m_wndListCtrl.GetItemData(iItem);
          iItem = m_wndListCtrl.GetNextItem(iItem, LVIS_SELECTED);
       }
 	   CDialog::OnOK();
